Move over final prompt methods
